# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Hello Peer是一款基于兴趣的社交应用。",
        "title": "Hello Peer API",
        "termsOfService": "API文档仅用于Hello Peer研发使用。",
        "contact": {
            "name": "Frank Sheng",
            "email": "726569998@qq.com"
        },
        "version": "0.1"
    },
    "host": "https://hellopeer.net",
    "basePath": "/api/v0.1",
    "paths": {
...
        "/images/{file_name}": {
            "get": {
                "security": [
                    {
                        "x-token": []
                    }
                ],
                "description": "获取图片",
                "tags": [
                    "Util"
                ],
                "summary": "获取图片",
                "operationId": "get_image",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Authorization",
                        "name": "x-token",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "0": {
                        "description": ""
                    }
                }
            }
        },
...
        "/user/deleteUserImage": {
            "post": {
                "security": [
                    {
                        "x-token": []
                    }
                ],
                "description": "删除用户图片",
                "tags": [
                    "User"
                ],
                "summary": "删除用户图片",
                "operationId": "delete_user_image",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Authorization",
                        "name": "x-token",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "文件名",
                        "name": "filename",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "0": {
                        "description": ""
                    }
                }
            }
        },
        "/user/getRandomUserDetails": {
            "get": {
                "security": [
                    {
                        "x-token": []
                    }
                ],
                "description": "可通过用户ID获取用户详情",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "获取广场用户详情",
                "operationId": "get_random_users_details",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Authorization",
                        "name": "x-token",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "用户ID",
                        "name": "uid",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "0": {
                        "description": "",
                        "schema": {
                            "$ref": "#/definitions/controllers.UserDetails"
                        }
                    }
                }
            }
        },
        "/user/getRandomUsers": {
            "post": {
                "security": [
                    {
                        "x-token": []
                    }
                ],
                "description": "可通过用户ID以及用户广场设置项获取随机用户列表",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "获取用户广场随机用户列表",
                "operationId": "get_random_users_list_in_square",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Authorization",
                        "name": "x-token",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "分页",
                        "name": "pagination",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Pagination"
                        }
                    }
                ],
                "responses": {
                    "0": {
                        "description": "",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/controllers.RandomUser"
                            }
                        }
                    }
                }
            }
        },
        "/user/getRecommendedUsers": {
            "get": {
                "security": [
                    {
                        "x-token": []
                    }
                ],
                "description": "可通过用户ID获取推荐用户列表",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "获取用户推荐用户列表",
                "operationId": "get_recommended_users",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Authorization",
                        "name": "x-token",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "0": {
                        "description": "",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/controllers.RecommendedUser"
                            }
                        }
                    }
                }
            }
        },
        "/user/getUserDetails": {
            "get": {
                "security": [
                    {
                        "x-token": []
                    }
                ],
                "description": "可根据用户ID获取用户详情",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "根据用户ID获取用户详情",
                "operationId": "get_user_details_by_id",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Authorization",
                        "name": "x-token",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "user_id",
                        "name": "uid",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "0": {
                        "description": "",
                        "schema": {
                            "$ref": "#/definitions/controllers.UserDetails"
                        }
                    }
                }
            }
        },
        "/user/getUserHomepageInfo": {
            "get": {
                "security": [
                    {
                        "x-token": []
                    }
                ],
                "description": "可获取用户首页信息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "获取用户首页信息",
                "operationId": "get_user_homepage_info",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Authorization",
                        "name": "x-token",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "0": {
                        "description": "",
                        "schema": {
                            "$ref": "#/definitions/models.HomepageInfo"
                        }
                    }
                }
            }
        },
        "/user/getUserInfoCompleteLevel": {
            "get": {
                "security": [
                    {
                        "x-token": []
                    }
                ],
                "description": "可根据用户ID获取用户信息完成度 0-未完成 1-基本信息 2-基本信息+头像",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "根据用户ID获取用户信息完成度",
                "operationId": "get_user_info_complete_level",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Authorization",
                        "name": "x-token",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "0": {
                        "description": "",
                        "schema": {
                            "$ref": "#/definitions/models.InfoCompletionLevel"
                        }
                    }
                }
            }
        },
...
        "/user/setUserBasicInfo": {
            "post": {
                "security": [
                    {
                        "x-token": []
                    }
                ],
                "description": "可通过用户token设置用户基础信息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "设置用户基础信息",
                "operationId": "set_user_basic_info",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Authorization",
                        "name": "x-token",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "用户基础信息",
                        "name": "basicInfo",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.BasicInfo"
                        }
                    }
                ],
                "responses": {
                    "0": {
                        "description": ""
                    }
                }
            }
        },
        "/user/setUserDetails": {
            "post": {
                "security": [
                    {
                        "x-token": []
                    }
                ],
                "description": "用户可设置信息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "设置用户信息",
                "operationId": "set_user_details",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Authorization",
                        "name": "x-token",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "用户信息",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controllers.UserDetailsUpdate"
                        }
                    }
                ],
                "responses": {
                    "0": {
                        "description": ""
                    }
                }
            }
        },
        "/user/setUserGender": {
            "post": {
                "security": [
                    {
                        "x-token": []
                    }
                ],
                "description": "可通过用户token设置用户性别",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "设置用户性别",
                "operationId": "set_user_gender",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Authorization",
                        "name": "x-token",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "用户性别",
                        "name": "gender",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.UserGender"
                        }
                    }
                ],
                "responses": {
                    "0": {
                        "description": ""
                    }
                }
            }
        },
...
        "/user/upload/setUserAvatar": {
            "post": {
                "security": [
                    {
                        "x-token": []
                    }
                ],
                "description": "用户可设置头像",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "设置用户头像",
                "operationId": "set_user_avatar",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Authorization",
                        "name": "x-token",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "file",
                        "description": "头像文件",
                        "name": "content",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "0": {
                        "description": ""
                    }
                }
            }
        },
        "/user/upload/setUserCover": {
            "post": {
                "security": [
                    {
                        "x-token": []
                    }
                ],
                "description": "用户可设置封面图",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "设置用户封面图",
                "operationId": "set_user_cover",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Authorization",
                        "name": "x-token",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "file",
                        "description": "封面文件",
                        "name": "content",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "0": {
                        "description": ""
                    }
                }
            }
        },
        "/user/upload/setUserImage": {
            "post": {
                "security": [
                    {
                        "x-token": []
                    }
                ],
                "description": "用户可上传图片",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "设置用户图片",
                "operationId": "set_user_image",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Authorization",
                        "name": "x-token",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "file",
                        "description": "用户图片",
                        "name": "content",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "0": {
                        "description": ""
                    }
                }
            }
        }
    },
    "definitions": {
        "controllers.RandomUser": {
            "type": "object",
            "properties": {
                "age": {
                    "type": "integer",
                    "example": 25
                },
                "coverImageUrl": {
                    "type": "string",
                    "example": "cover.png"
                },
                "lat": {
                    "type": "number",
                    "example": 22.51
                },
                "lng": {
                    "type": "number",
                    "example": 113.95
                },
                "location": {
                    "type": "string",
                    "example": "南山区"
                },
                "occupation": {
                    "type": "string",
                    "example": "平面设计师"
                },
                "petName": {
                    "type": "string",
                    "example": "Amy"
                },
                "userName": {
                    "type": "string",
                    "example": "豆豆"
                }
            }
        },
        "controllers.RecommendedUser": {
            "type": "object",
            "properties": {
                "age": {
                    "type": "integer",
                    "example": 25
                },
                "coverImageUrl": {
                    "type": "string",
                    "example": "cover.png"
                },
                "images": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "img1.jpg",
                        "img2.jpg"
                    ]
                },
                "lat": {
                    "type": "number",
                    "example": 22.51
                },
                "lng": {
                    "type": "number",
                    "example": 113.95
                },
                "location": {
                    "type": "string",
                    "example": "南山区"
                },
                "occupation": {
                    "type": "string",
                    "example": "平面设计师"
                },
                "petName": {
                    "type": "string",
                    "example": "Amy"
                },
                "tags": {
                    "type": "string",
                    "example": "猫控 读书达人 电影爱好者"
                },
                "userName": {
                    "type": "string",
                    "example": "豆豆"
                },
                "verified": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
        "controllers.UserDetails": {
            "type": "object",
            "properties": {
                "age": {
                    "type": "integer",
                    "example": 25
                },
                "birthday": {
                    "type": "string",
                    "example": "1988-10-2"
                },
                "constellation": {
                    "type": "string",
                    "example": "处女座"
                },
                "cover_Image": {
                    "type": "string",
                    "example": "cover.png"
                },
                "declaration": {
                    "type": "string",
                    "example": "交友宣言"
                },
                "education": {
                    "type": "string",
                    "example": "本科"
                },
                "gender": {
                    "type": "integer",
                    "example": 1
                },
                "height": {
                    "type": "string",
                    "example": "165cm"
                },
                "hobbies": {
                    "type": "string",
                    "example": "兴趣爱好"
                },
                "hometown": {
                    "type": "string",
                    "example": "湖南长沙"
                },
                "images": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "img1.jpg",
                        "img2.jpg"
                    ]
                },
                "income": {
                    "type": "string",
                    "example": "15-25w"
                },
                "location": {
                    "type": "string",
                    "example": "深圳"
                },
                "marriage": {
                    "type": "string",
                    "example": "未婚"
                },
                "occupation": {
                    "type": "string",
                    "example": "平面设计师"
                },
                "purpose": {
                    "type": "string",
                    "example": "交友"
                },
                "self_Desc": {
                    "type": "string",
                    "example": "自我描述"
                },
                "tags": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "猫控",
                        "读书达人",
                        "电影爱好者",
                        "旅行者"
                    ]
                },
                "the_One": {
                    "type": "string",
                    "example": "希望另一半的样子"
                },
                "userName": {
                    "type": "string",
                    "example": "豆豆"
                },
                "weight": {
                    "type": "string",
                    "example": "43kg"
                }
            }
        },
...
        "models.HomepageInfo": {
            "type": "object",
            "properties": {
                "avatar": {
                    "type": "string",
                    "example": "avatar1-1234.jpg"
                },
                "location": {
                    "type": "string",
                    "example": "广东,深圳"
                },
                "pet_food": {
                    "type": "integer",
                    "example": 320
                },
                "pet_name": {
                    "type": "string",
                    "example": "宠物猫猫"
                },
                "stat": {
                    "$ref": "#/definitions/models.RelationStat"
                },
                "user_name": {
                    "type": "string",
                    "example": "用户1"
                }
            }
        },
        "models.InfoCompletionLevel": {
            "type": "object",
            "properties": {
                "level": {
                    "type": "integer",
                    "example": 1
                }
            }
        },
        "models.Pagination": {
            "type": "object",
            "properties": {
                "limit": {
                    "type": "integer",
                    "example": 4
                },
                "offset": {
                    "type": "integer",
                    "example": 0
                }
            }
        },
...
        "models.UserGender": {
            "type": "object",
            "properties": {
                "gender": {
                    "type": "integer",
                    "example": 1
                }
            }
        },
...
    }
}